funscript: Refactor calcuation functions
